+++ AdventureBackpacks/Properties/AssemblyInfo.cs
@@ -32,8 +32,8 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 #if ! API
-[assembly: AssemblyVersion("1.9.13.4")]
-[assembly: AssemblyFileVersion("1.9.13.4")]
+[assembly: AssemblyVersion("1.9.13.5")]
+[assembly: AssemblyFileVersion("1.9.13.5")]
 #else
 [assembly: AssemblyVersion("1.2.0")]
 [assembly: AssemblyFileVersion("1.2.0")]